requirements.txt
Django==3.0.6
google-cloud-datastore==1.12.0
google-cloud-ndb==1.2.1
google-auth==1.15.0
rauth
